refactor: updates file names due to file naming convention-consistency
